+++ public/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.9.0] - 2025-07-14
+
+### Changed
+- Switched all audio normalization from RMS to peak normalization for more consistent results
+- Transparent limiter is now always applied by default to prevent clipping in all exported audio
+
+### Fixed
+- Prevented all possible output clipping by enforcing limiter in the audio processing pipeline
+
 ## [0.8.0] - 2025-07-11
 
 ### Added